Add TouchJoystick!!
Change texture for asteroid and ship.

Fix resolution changes in android. (And touch position)

Stats and settings are saved/loaded correctly on Android.

Full Android compatibility (?) let'sfGOOOO

+++ include/utils.hpp
@@ -50,7 +50,7 @@ enum RotationPoint : uint16_t {
 SDL_FPoint getPivotFromRotationPoint(RotationPoint rp, int texW, int texH);
 
 inline float randFloat(float min, float max) {
-    return min + static_cast<float>(rand()) / RAND_MAX * (max - min);
+	return min + static_cast<float>(rand()) / static_cast<float>(RAND_MAX) * (max - min);
 }
 
 std::string normalizePath(const std::string& path);
